feat : secondPage survey option controller 수정
Refs:[#60]
